% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocknr_to_hash.R
\name{blocknr_to_hash}
\alias{blocknr_to_hash}
\title{blocknr_to_hash}
\usage{
blocknr_to_hash(blocknr, writecsv = FALSE, filename)
}
\arguments{
\item{blocknr}{a vector of blocknumbers}

\item{writecsv}{logical, should a .csv file be created}

\item{filename}{optional: the name of the file to be created (has to be "xxx.csv")}
}
\value{
a data.frame with the blocknr and the blockhash256 character long hash of the blocknr
}
\description{
Gets the blockhash corresponding to a certain blocknr from the rawmempool
}
\examples{
\dontrun{
blocknr_to_hash(blocknr = 500000:500010)
blocknr_to_hash(blocknr = 500000:500010, writecsv = TRUE)
blocknr_to_hash(blocknr = 500000:500010, writecsv = TRUE, filename = "hash.csv")
}
}
