% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_duplicates_nonvaltx.R
\name{eliminate_duplicates_nonvaltx}
\alias{eliminate_duplicates_nonvaltx}
\title{eliminate_duplicates_nonvaltx}
\usage{
eliminate_duplicates_nonvaltx(blocknr)
}
\arguments{
\item{blocknr}{vector with blocknumbers for which to create the files without duplicates in an extra folder}
}
\description{
eliminate_duplicates_nonvaltx
}
\examples{
\dontrun{
eliminate_duplicates_nonvaltx(500000:501000)
}
}
