% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rawmempool.R
\name{fill_missing_rawmempool}
\alias{fill_missing_rawmempool}
\title{fill_missing_rawmempool}
\usage{
fill_missing_rawmempool(blocknr)
}
\arguments{
\item{blocknr}{sequence for which missing rawmempool files will be created}
}
\description{
Since we are only taking a snapshot of the rawmempool every minute it can happpen
that two blocks get verified inbetween two snapshots, thus we are missing rawmempoolfiles.
To coutneract this we look at the rawmempool of the block before and dismiss the
validated transactions.
}
\examples{
\dontrun{
fill_missing_rawmempool(blocknr = 500000:501000)
}
}
