% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blockinfo.R
\name{get_blockinfo}
\alias{get_blockinfo}
\title{get_blockinfo}
\usage{
get_blockinfo(blocknr, blockhash, writecsv = FALSE, filename)
}
\arguments{
\item{blocknr}{vector of blocknumbers for which to generate the blockinfo file.}

\item{blockhash}{vector of blockhashes corresponding to the blocknr.}

\item{writecsv}{logical, should a .csv file be created.}

\item{filename}{optional: the name of the file to be created (has to be "xxx.csv").}
}
\value{
a data.frame with the infos of the requested blocks.
}
\description{
Fetch basic blockinfos fo a vector of blockhashes and optionally save them into a .csv file
}
\details{
For eash blocknumber one line will be added to the data.frame. It contains the blocknumber,
 blocksize in bytes, blockweight, blocktime, difficulty, number of transactions
 in the rawmempool, number of verified transaction in this block and the time
}
\examples{
\dontrun{
blocknr <- 499999
blockhash <- "0000000000000000007962066dcd6675830883516bcf40047d42740a85eb2919"
get_blockinfo(blocknr = blocknr, blockhash = blockhash)
get_blockinfo(blocknr = blocknr, blockhash = blockhash, writecsv = TRUE)
get_blockinfo(blocknr = blocknr, blockhash = blockhash, writecsv = TRUE, filename = "hashes.csv")
}
}
