% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawmempool_functions.R
\name{get_fee_recommendation}
\alias{get_fee_recommendation}
\title{get_fee_recommendation}
\usage{
get_fee_recommendation(blocks)
}
\arguments{
\item{blocks}{required number of blocks}
}
\value{
recommended fees
}
\description{
get_fee_recommendation
}
