% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blocknr.R
\name{get_latest_blocknr}
\alias{get_latest_blocknr}
\title{get_latest_blocknr}
\usage{
get_latest_blocknr()
}
\value{
a single number of the latest block
}
\description{
Calls the chainquery.com site and fetches the the number of the latest block of the blockchain
}
\examples{
get_latest_blocknr()
}
