% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonvaltx.R
\name{get_nonvaltx}
\alias{get_nonvaltx}
\title{get_nonvaltx}
\usage{
get_nonvaltx(blocknr, writecsv = FALSE, filename)
}
\arguments{
\item{blocknr}{vector of blocknr for which to create .csv files containing the nonvalidated
transactions}

\item{writecsv}{logical, should a .csv file be created}

\item{filename}{optional: the name of the file(s) to be created (has to be "xxx.csv")}
}
\description{
Creates csv files for the blocks in blocknr. In order to do so the rawmempool files and the validated_tx files have to exist, because the nonvalidated_tx are created by dismissing the validated_tx from the rawmempool
}
\examples{
\dontrun{
get_nonvaltx(491464:491465, writecsv = TRUE)
}
}
