% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawmempool_functions.R
\name{get_raw_mempool}
\alias{get_raw_mempool}
\title{get_raw_mempool}
\usage{
get_raw_mempool()
}
\value{
a data.frame with the blocknr and the blockhash256 character long hash of the blocknr
}
\description{
get_raw_mempool
}
