% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rawmempool.R
\name{get_rawmempool}
\alias{get_rawmempool}
\title{get_rawmempool}
\usage{
get_rawmempool(writecsv = FALSE, filename)
}
\arguments{
\item{writecsv}{logical, should a .csv file be created}

\item{filename}{optional: the name of the file to be created (has to be "xxx.csv")}
}
\value{
a data.frame with size, fee, time, height, and transactionid or NULL if a .csv file was saved.
}
\description{
Calls the chainquery.com site and makes a snapshot of the transactions of the rawmempool.
}
\details{
Since it needs to download a lot of data this function can take up to 15 sec. to run.
A folder named "data" must exist in the current work directory if writecsv is set to TRUE.
The default name is rawmempool_####.csv where #### is the next blocknumber.
By default it overwrites itself if there is already a .csv file of the transaction for the next block.
This guarantees that only one file per block exists which is the most recent one.
}
\examples{
\dontrun{
get_rawmempool()
get_rawmempool(writecsv = TRUE)
get_rawmempool(writecsv = TRUE, filename = "snapshot01.csv")
}
}
