% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawmempool_functions.R
\name{get_transaction_data}
\alias{get_transaction_data}
\title{get_transaction_data}
\usage{
get_transaction_data(tc_id)
}
\arguments{
\item{tc_id}{transaction id}
}
\value{
transaction_data
}
\description{
get_transaction_data
}
