% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/merge_rmp_valtx_nonvaltxnoduplicates_blockinfo.R
\name{merge_rmp_valtx_nonvaltxnoduplicates_blockinfo}
\alias{merge_rmp_valtx_nonvaltxnoduplicates_blockinfo}
\title{merge_rmp_valtx_nonvaltxnoduplicates_blockinfo}
\usage{
merge_rmp_valtx_nonvaltxnoduplicates_blockinfo(blocknr, blockhash, blockinfo,
  writecsv = FALSE)
}
\arguments{
\item{blocknr}{vector of blocknumbers for which to generate the merged files.}

\item{blockhash}{vector of blockhashes corresponding to the blocknr.}

\item{blockinfo}{data.frame, which contains data of the single blocks to be merged, created by get_blockinfo()}

\item{writecsv}{logical, should a .csv file be created.}
}
\value{
returns a dataframe of the merged info where every transaction is included only once

If writecsv = TRUE it creates a .csv with the following columns:blocknr, validated, time, height, fee, size, blocksize, blockweight, blocktime, difficulty, nr_tx_rmp, nr_tx_vtx, file_creation_time, nr_obs_valtx
}
\description{
Merges the infos of the rawmempool with the validated and nonvalidated transactions without dupicates and the blockinfo.
This means for the whole data.frame which consists out of many block each transaction will only be includes once.
}
\examples{
\dontrun{
merge_rmp_valtx_nonvaltxnoduplicates_blockinfo(blocknr = blocknr,
blockhash = blockhash, writecsv = TRUE)
}
}
