% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawmempool_functions.R
\name{recomended_fees}
\alias{recomended_fees}
\title{get_fee_recommendation}
\usage{
recomended_fees(blocks)
}
\arguments{
\item{blocks}{required number of blocks}
}
\value{
recommended fees
}
\description{
get_fee_recommendation
}
