% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folder.R
\name{setup_folder}
\alias{setup_folder}
\title{Setup Folders}
\usage{
setup_folder()
}
\description{
creates the folders necessary for the data that will be produced. The rawmempool
folder has to be filled with the raw webscraped data since it will be used as the basis
for the data generation.
}
\examples{
setup_folder()
}
