% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valtx.R
\name{get_valtx}
\alias{get_valtx}
\title{get_valtx}
\usage{
get_valtx(blocknr, writecsv = FALSE)
}
\arguments{
\item{blocknr}{vector of blocknumbers}

\item{writecsv}{logical, should a .csv file be created}
}
\description{
Create a .csv file for each blocknr which contains the transaction_id and the blocknumber
}
